cdn_plus/requirements.txt
asgiref==3.7.2
boto3==1.33.12
botocore==1.33.12
Django==5.0
django-mapbox-location-field==2.0.0
django-webpack-loader==2.0.1
djangorestframework==3.14.0
jmespath==1.0.1
leaflet==0.0.3
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
